Wire Frames Photos
Instead of making an online Frame, I decided to make it by pen and paper. I wouldn't suggest this, its messy and hard to read. however it can help when making such a large site to draw it up this way to help organize the skeleton of your site. Ive added these to get an idea of how I went about it. The client would never see such a rough copy. However they would receive a completed power point with additional information, this is only for the developers eyes.